added template for JPA

+++ spring-14/src/main/java/ru/otus/hw/services/BookServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.otus.hw.converters.BookConverter;
 import ru.otus.hw.dto.BookDto;
+import ru.otus.hw.dto.BookFormDto;
 import ru.otus.hw.exceptions.EntityNotFoundException;
 import ru.otus.hw.models.Book;
 import ru.otus.hw.repositories.AuthorRepository;
@@ -12,7 +13,6 @@ import ru.otus.hw.repositories.BookRepository;
 import ru.otus.hw.repositories.GenreRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -27,42 +27,53 @@ public class BookServiceImpl implements BookService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<BookDto> findById(long id) {
-        return bookRepository.findById(id).map(bookConverter::fromDomainObject);
+    public BookDto findById(long id) {
+        return bookConverter.fromDomainObject(getBook(id));
     }
 
     @Override
     @Transactional(readOnly = true)
     public List<BookDto> findAll() {
-        return bookRepository.findAll().stream()
+        return bookRepository.findAll()
+                .stream()
                 .map(bookConverter::fromDomainObject)
                 .toList();
     }
 
     @Override
     @Transactional
-    public BookDto insert(String title, long authorId, long genreId) {
-        return bookConverter.fromDomainObject(save(0, title, authorId, genreId));
+    public BookDto insert(BookFormDto bookFormDto) {
+        return bookConverter.fromDomainObject(save(0, bookFormDto));
     }
 
     @Override
     @Transactional
-    public BookDto update(long id, String title, long authorId, long genreId) {
-        return bookConverter.fromDomainObject(save(id, title, authorId, genreId));
+    public BookDto update(long id, BookFormDto bookFormDto) {
+        return bookConverter.fromDomainObject(save(id, bookFormDto));
     }
 
     @Override
     @Transactional
     public void deleteById(long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Book with id %d not found".formatted(id));
+        }
         bookRepository.deleteById(id);
     }
 
-    private Book save(long id, String title, long authorId, long genreId) {
-        var author = authorRepository.findById(authorId)
-                .orElseThrow(() -> new EntityNotFoundException("Author with id %d not found".formatted(authorId)));
-        var genre = genreRepository.findById(genreId)
-                .orElseThrow(() -> new EntityNotFoundException("Genre with id %d not found".formatted(genreId)));
-        var book = new Book(id, title, author, genre);
+    private Book getBook(long id) {
+        return bookRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Book with id %d not found".formatted(id)));
+    }
+
+    private Book save(long id, BookFormDto bookFormDto) {
+        var author = authorRepository.findById(bookFormDto.authorId())
+                .orElseThrow(() ->
+                        new EntityNotFoundException("Author with id %d not found".formatted(bookFormDto.authorId())));
+        var genre = genreRepository.findById(bookFormDto.genreId())
+                .orElseThrow(() ->
+                        new EntityNotFoundException("Genre with id %d not found".formatted(bookFormDto.genreId())));
+        var book = new Book(id, bookFormDto.title(), author, genre);
         return bookRepository.save(book);
     }
 }
